FEAT : 룸 현황 조회 기능 추가

+++ ReservationSystem/src/ReservationSystem/Guest.java
@@ -126,7 +126,7 @@ public class Guest {
 			Thread.sleep(1500);
 			System.out.println("취소수수료가 10% 부과되어 환불됩니다.");
 			Thread.sleep(1500);
-			System.out.println("환불 금액은 " + (money * 0.9) + "원 입니다.");
+			System.out.println("환불 금액은 " + (int)(money * 0.9) + "원 입니다.");
 			Thread.sleep(1500);
 		} catch (InterruptedException e) {
 			e.printStackTrace();
@@ -136,15 +136,14 @@ public class Guest {
 		System.out.print("Y/N : ");
 		String ny = sc.next();
 		if (ny.charAt(0) == 'Y' || ny.charAt(0) == 'y') {
-			hotelmoney -= (int)(money * 0.9);
-			name="";
+			hotelmoney -= (int) (money * 0.9);
+			name = "";
 			map1.put(room, false);
 			System.out.println();
 			System.out.println(room + "호실의 예약 취소가 완료되었습니다.");
 			System.out.println("감사합니다.");
-			
 
-		}else if (ny.charAt(0) == 'N' || ny.charAt(0) == 'n') {
+		} else if (ny.charAt(0) == 'N' || ny.charAt(0) == 'n') {
 			System.out.println();
 			System.out.println("고객님 요청으로 예약이 취소되지 않았습니다.");
 			System.out.println("감사합니다.");
@@ -152,20 +151,21 @@ public class Guest {
 
 		return hotelmoney;
 	}
+
 	public int roomReservationChange(Scanner sc, int hotelmoney) {
 		System.out.println();
 		System.out.println("고객님의 예약 정보 입니다.");
 		System.out.println();
-		
+
 		System.out.println("숙박 일수 : " + day + "박 " + (day + 1) + "일");
 		System.out.println("체크인 : " + year + "년 " + month + "월 " + date + "일");
 		System.out.println("체크아웃 : " + year + "년 " + month + "월 " + (date + day) + "일");
-		
+
 		System.out.print("몇 박을 연장하시겠습니까? : ");
-		int day_c=sc.nextInt();
-		int money_c=55000*day_c;
-		day+=day_c;
-		
+		int day_c = sc.nextInt();
+		int money_c = 55000 * day_c;
+		day += day_c;
+
 		System.out.println();
 		System.out.println(day_c + "박 추가 결제 비용은 " + money_c + "원 입니다.");
 		try {
@@ -175,26 +175,24 @@ public class Guest {
 		}
 		System.out.println();
 		System.out.println("결제되었습니다. 감사합니다.");
-		
+
 		try {
 			Thread.sleep(1500);
 		} catch (InterruptedException e) {
 			e.printStackTrace();
 		}
-		
+
 		System.out.println();
 		System.out.println("변경된 예약정보 입니다.");
 		System.out.println();
 		System.out.println("숙박 일수 : " + day + "박 " + (day + 1) + "일");
 		System.out.println("체크인 : " + year + "년 " + month + "월 " + date + "일");
 		System.out.println("체크아웃 : " + year + "년 " + month + "월 " + (date + day) + "일");
-		
-		
+
 		hotelmoney += money_c;
 
 		return hotelmoney;
 
-		
 	}
 
 }
